+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 -   Fix rank tooltip on perfcards in CP performances page (#108)
 -   Fix outdated refresh route in withRetry (#112)
+-   Fix users management link in menu (#113)
 
 ## [0.34.0] - 2022-09-13